Modify ids types from int to Long

+++ src/main/java/com/vroom/vroom/model/User.java
@@ -27,7 +27,7 @@ public class User {
         this.isAdmin = isAdmin;
     }
 
-    public int getIdUser() {
+    public Long getIdUser() {
         return idUser;
     }
